Consolidate Webpack config options into a single variable
Updated the call to webpackStream() in the javascript task. Replaced the object being passed as the first argument with the variable webpackConfig. Having been previously initialized, it as been updated to include the properties that were removed from the aforementioned argument. The goal is to consolidate the configuration into a single location so that expanding on it with additional properties can be accomplished more easily.