Restore the template in working state
The November 2017 commits starting with 55b5b0b1 used the template repository
to inject code for the a customer project.

The commit ff93bd892 hided this situation fixing the README.

This change fixes this situation restoring genuine Foundation settings,
the former style guide and content, without the customer customisation.

Documentation and supports links are updated for Foundation 6.